Expose skybridge typegen command for prepare hook
